Refactor: Update SendSmsCodeButton state in LoginScreen
This commit refactors the `SendSmsCodeButton` in `LoginScreen.kt`.

The `enabled` state of the `TextButton` within `SendSmsCodeButton` is now determined by both the countdown status and the `SendSmsCodeUiState`. The button will be disabled if a countdown is in progress OR if the `sendSmsCodeState` is `SendSmsCodeUiState.Loading`.

This change ensures the "Send Code" button is appropriately disabled while an SMS code is being sent, in addition to during the countdown period.